Restore the type of NVM offset from 32bit to 16bit

+++ common/c_callbacks.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 struct ParameterInfo;
 // FIXME: descript in common/platform.cpp
-typedef void (*data_preservation_func)(struct ParameterInfo *param, uint32_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
+typedef void (*data_preservation_func)(struct ParameterInfo *param, uint16_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
 void record_overflow_handling_overhead(uint32_t cycles);
 #ifdef __cplusplus
 }
