+++ common/cnn_common.cpp
@@ -78,7 +78,7 @@ const uint8_t* get_param_first_tile_index_base_pointer(const ParameterInfo *para
 #endif
 
 // FIXME: descript in common/platform.cpp
-int16_t get_q15_param(Model* model, const ParameterInfo *param, uint32_t i) {
+int16_t get_q15_param(Model* model, const ParameterInfo *param, uint16_t i) {
     MY_ASSERT(param->bitwidth == 16);
     if (param->slot == SLOT_TEST_SET) {
         int16_t ret;
@@ -230,7 +230,7 @@ uint16_t get_col_val(Model *model, const ParameterInfo *params, uint16_t col_ind
 #endif // SPARSE
 
 // FIXME: descript in common/platform.cpp
-void my_memcpy_from_param(Model* model, void *dest, const ParameterInfo *param, uint32_t offset_in_word, size_t n) {
+void my_memcpy_from_param(Model* model, void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n) {
     if (param->slot == SLOT_TEST_SET) {
         read_from_samples(dest, offset_in_word, n);
     } else if (param->slot >= SLOT_CONSTANTS_MIN) {
