+++ common/my_dsplib.h
@@ -9,11 +9,11 @@ void my_fill_q15(int16_t value, int16_t *pDst, uint32_t blockSize);
 void my_offset_q15(const int16_t *pSrc, int16_t offset, int16_t *pDst, uint32_t blockSize);
 // FIXME: common/platform.cpp
 void my_matrix_mpy_q15(uint16_t A_rows, uint16_t A_cols, uint16_t B_rows, uint16_t B_cols, int16_t *pSrcA, int16_t *pSrcB, int16_t *pDst,
-                       ParameterInfo *param, uint32_t offset_in_word, size_t values_to_preserve,
+                       ParameterInfo *param, uint16_t offset_in_word, size_t values_to_preserve,
                        uint16_t mask, int16_t n_keep_state_bits);
 // FIXME: common/platform.cpp
 void my_matrix_mpy_q15_to_vm(uint16_t A_rows, uint16_t A_cols, uint16_t B_rows, uint16_t B_cols, int16_t *pSrcA, int16_t *pSrcB, int16_t *pDst,
-                       ParameterInfo *param, uint32_t offset_in_word, size_t values_to_preserve,
+                       ParameterInfo *param, uint16_t offset_in_word, size_t values_to_preserve,
                        uint16_t mask, int16_t n_keep_state_bits);
 void my_max_q15(const int16_t *pSrc, uint32_t blockSize, int16_t *pResult, uint16_t *pIndex);
 void my_min_q15(const int16_t *pSrc, uint32_t blockSize, int16_t *pResult, uint16_t *pIndex);
